WIP: refactor integration test
