+++ contracts/hackatom/tests/integration.rs
@@ -1,19 +1,17 @@
 use std::fs;
 use std::str::from_utf8;
 
-use wasmer_runtime::{compile_with, Ctx, Func, func, imports};
-use wasmer_runtime_core::{Instance};
-use wasmer_clif_backend::CraneliftCompiler;
-
 use hackatom::contract::{RegenInitMsg};
 use cosmwasm::imports::Storage;
 use cosmwasm::types::{coin, mock_params};
 
 mod memory;
-mod context;
+mod exports;
+mod wasmer;
 
-use crate::context::{create_unmanaged_storage, destroy_unmanaged_storage, with_storage_from_context};
+use crate::exports::{do_read, do_write, setup_context};
 use crate::memory::{read_memory, write_memory, allocate};
+use crate::wasmer::{Func, func, imports, wasm_instance};
 
 /**
 This integration test tries to run and call the generated wasm.
@@ -24,8 +22,6 @@ cargo wasm && wasm-gc ./target/wasm32-unknown-unknown/release/hackatom.wasm
 Then running `cargo test` will validate we can properly call into that generated data.
 **/
 
-
-
 #[test]
 fn run_contract() {
     let wasm_file = "./target/wasm32-unknown-unknown/release/hackatom.wasm";
@@ -35,7 +31,7 @@ fn run_contract() {
     // TODO: set up proper callback for read and write here
     // TODO: figure out passing state
     let import_object = imports! {
-        || (create_unmanaged_storage(), destroy_unmanaged_storage),
+        || { setup_context() },
         "env" => {
             "c_read" => func!(do_read),
             "c_write" => func!(do_write),
@@ -43,8 +39,7 @@ fn run_contract() {
     };
 
     // create the instance
-    let module = compile_with(&wasm, &CraneliftCompiler::new()).unwrap();
-    let mut instance = module.instantiate (&import_object).unwrap();
+    let mut instance = wasm_instance(wasm, import_object);
 
     // prepare arguments
     let params = mock_params("creator", &coin("1000", "earth"), &[]);
@@ -71,23 +66,4 @@ fn run_contract() {
     let str_res = from_utf8(&res).unwrap();
     assert_eq!(str_res , "{\"msgs\":[]}");
 }
-
-/*** mocks to stub out actually db writes as extern "C" ***/
-
-fn do_read(ctx: &mut Ctx, key_ptr: i32, val_ptr: i32) -> i32 {
-    let key = read_memory(ctx, key_ptr);
-    let mut value: Option<Vec<u8>> = None;
-    with_storage_from_context(ctx, |store| value = store.get(&key));
-    match value {
-        Some(buf) => write_memory(ctx, val_ptr, &buf),
-        None => 0,
-    }
-}
-
-fn do_write(ctx: &mut Ctx, key: i32, value: i32) {
-    let key = read_memory(ctx, key);
-    let value = read_memory(ctx, value);
-    with_storage_from_context(ctx, |store| store.set(&key, &value));
-}
-
